MAFuture: Fix integer precision issues

+++ MAFuture/MACompoundFuture.m
@@ -31,8 +31,8 @@
     else if([sig methodReturnType][0] != @encode(id)[0]) return NO;
     
     // it exists, returns an object, but does it return any non-objects by reference?
-    unsigned num = [sig numberOfArguments];
-    for(unsigned i = 2; i < num; i++)
+    NSUInteger num = [sig numberOfArguments];
+    for(NSUInteger i = 2; i < num; i++)
     {
         const char *type = [sig getArgumentTypeAtIndex: i];
         
@@ -92,8 +92,8 @@
         _MALazyBlockFuture *invocationFuture = nil;
         NSMutableArray *parameterDatas = nil;
         NSMethodSignature *sig = [invocation methodSignature];
-        unsigned num = [sig numberOfArguments];
-        for(unsigned i = 2; i < num; i++)
+        NSUInteger num = [sig numberOfArguments];
+        for(NSUInteger i = 2; i < num; i++)
         {
             const char *type = [sig getArgumentTypeAtIndex: i];
             if(type[0] == '^' && type[1] == '@')
